+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## unreleased
 
 * Fixed: Selected fonts not loading ([#44](https://github.com/amur-tiger/fanfiction-enhancements/issues/44))
+* Changed: Icon for alerts changed from bookmark to bell ([#59](https://github.com/amur-tiger/fanfiction-enhancements/issues/59))
+* New Feature: Tooltips for Download/Follow/Favorite buttons ([#58](https://github.com/amur-tiger/fanfiction-enhancements/issues/58))
 
 ## 0.7.2